refactor(frontend): change name of module
