+++ apps/frontend/app/routes/_dashboard.fitness.workouts.$id._index.tsx
@@ -72,7 +72,7 @@ import {
 import {
 	duplicateOldWorkout,
 	getExerciseDetailsQuery,
-} from "~/lib/state/workout";
+} from "~/lib/state/fitness";
 import {
 	createToastHeaders,
 	getAuthorizationHeader,
